feat(snapshot): add GetOrCreateSnapshot endpoint with cached snapshot strategy
- Added POST /report/get-or-create-snapshot endpoint
- Integrated SnapshotApplicationService.GetOrCreateSnapshotAsync
- Enables snapshot reuse to avoid full recalculation
- Improves performance by calculating only delta when snapshot exists
- Keeps calculate-snapshot endpoint for full recalculation scenarios

+++ Report.API/Models/ViewModels/InventoryBalanceResultViewModel.cs
@@ -4,5 +4,5 @@ public class InventoryBalanceResultViewModel
 {
     public string ProdutoID { get; set; }
     public string DepositoID { get; set; }
-    public decimal Saldo { get; set; }
+    public double Saldo { get; set; }
 }
